Move ipp related classes to inperson package
