+++ example/src/androidTest/java/org/wordpress/android/fluxc/mocked/MockedStack_WCPayTest.kt
@@ -6,14 +6,14 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.CAPTURE_ERROR
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.MISSING_ORDER
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.SERVER_ERROR
-import org.wordpress.android.fluxc.model.payments.PaymentAccountResult.PaymentAccountStatusEnum
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationErrorType.GenericError
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationErrorType.InvalidPostalCode
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationErrorType.MissingAddress
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.CAPTURE_ERROR
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.MISSING_ORDER
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.SERVER_ERROR
+import org.wordpress.android.fluxc.model.payments.inperson.PaymentAccountResult.PaymentAccountStatusEnum
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationErrorType.GenericError
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationErrorType.InvalidPostalCode
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationErrorType.MissingAddress
 import org.wordpress.android.fluxc.module.ResponseMockingInterceptor
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.pay.InPersonPaymentsRestClient
 import javax.inject.Inject
