+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/payments/inperson/ConnectionTokenResult.kt
@@ -1,3 +1,3 @@
-package org.wordpress.android.fluxc.model.payments
+package org.wordpress.android.fluxc.model.payments.inperson
 
 data class ConnectionTokenResult(val token: String, val isTestMode: Boolean)
