+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/payments/inperson/CreateCustomerByOrderIdResult.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.fluxc.model.payments
+package org.wordpress.android.fluxc.model.payments.inperson
 
 import com.google.gson.annotations.SerializedName
