+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/network/rest/wpcom/wc/pay/InPersonPaymentsRestClient.kt
@@ -5,20 +5,20 @@ import com.android.volley.RequestQueue
 import org.wordpress.android.fluxc.Dispatcher
 import org.wordpress.android.fluxc.generated.endpoint.WOOCOMMERCE
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.payments.CapturePaymentError
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.CAPTURE_ERROR
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.GENERIC_ERROR
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.MISSING_ORDER
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.NETWORK_ERROR
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
-import org.wordpress.android.fluxc.model.payments.CapturePaymentErrorType.SERVER_ERROR
-import org.wordpress.android.fluxc.model.payments.CapturePaymentResponsePayload
-import org.wordpress.android.fluxc.model.payments.PaymentAccountResult
-import org.wordpress.android.fluxc.model.payments.CreateCustomerByOrderIdResult
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationError
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationErrorType
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationResult
-import org.wordpress.android.fluxc.model.payments.TerminalStoreLocationResult.StoreAddress
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentError
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.CAPTURE_ERROR
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.GENERIC_ERROR
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.MISSING_ORDER
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.NETWORK_ERROR
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentErrorType.SERVER_ERROR
+import org.wordpress.android.fluxc.model.payments.inperson.CapturePaymentResponsePayload
+import org.wordpress.android.fluxc.model.payments.inperson.PaymentAccountResult
+import org.wordpress.android.fluxc.model.payments.inperson.CreateCustomerByOrderIdResult
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationError
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationErrorType
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationResult
+import org.wordpress.android.fluxc.model.payments.inperson.TerminalStoreLocationResult.StoreAddress
 import org.wordpress.android.fluxc.network.BaseRequest.GenericErrorType
 import org.wordpress.android.fluxc.network.UserAgent
 import org.wordpress.android.fluxc.network.rest.wpcom.BaseWPComRestClient
